user authentication done

+++ django_project/doc_appoint/role/views.py
@@ -7,20 +7,24 @@ from django.core.mail import send_mail
 
 
 def index(request):
-    duplicate_entries = Role.objects.filter(id='1')
-    duplicate_entries.delete()
-    send_mail("write you subject","How are you?",'salmanmdsmdsultan92@gmail.com',['aminulmahi12@gmail.com'])
-    all_data = Role.objects.all()
-    if(len(all_data)==0):
-        status = False
-    else:
-        status = True
-    
-    msg = messages.get_messages(request)
-    print(type(msg))
+    if 'user_id' in request.session:
 
-    data ={'all_data':all_data,'status':status,'msg':msg}
-    return render(request,'admin/role.html',data)
+        duplicate_entries = Role.objects.filter(id='1')
+        duplicate_entries.delete()
+        send_mail("write you subject","How are you?",'salmanmdsmdsultan92@gmail.com',['aminulmahi12@gmail.com'])
+        all_data = Role.objects.all()
+        if(len(all_data)==0):
+            status = False
+        else:
+            status = True
+        
+        msg = messages.get_messages(request)
+        print(type(msg))
+        print(request.session.get('user_name'))
+
+        data ={'all_data':all_data,'status':status,'msg':msg,'name':request.session.get('user_name')}
+        return render(request,'admin/role.html',data)
+    return redirect('login')
 
 def insert(request):
     role = request.POST.get('role')
